move fixtures which should be potentially common to util module
